+++ include/okapi/api.hpp
@@ -21,6 +21,8 @@
 #include "okapi/api/control/async/asyncVelIntegratedController.hpp"
 #include "okapi/api/control/controllerInput.hpp"
 #include "okapi/api/control/controllerOutput.hpp"
+#include "okapi/api/control/iterative/iterativePosPidController.hpp"
+#include "okapi/api/control/iterative/iterativeVelPidController.hpp"
 #include "okapi/api/control/util/flywheelSimulator.hpp"
 #include "okapi/api/control/util/settledUtil.hpp"
 #include "okapi/impl/control/async/asyncControllerFactory.hpp"
@@ -28,8 +30,6 @@
 #include "okapi/impl/control/async/asyncVelPidController.hpp"
 #include "okapi/impl/control/async/asyncWrapper.hpp"
 #include "okapi/impl/control/iterative/iterativeMotorVelocityController.hpp"
-#include "okapi/impl/control/iterative/iterativePosPidController.hpp"
-#include "okapi/impl/control/iterative/iterativeVelPidController.hpp"
 #include "okapi/impl/control/util/controllerRunner.hpp"
 #include "okapi/impl/control/util/pidTuner.hpp"
 #include "okapi/impl/control/util/settledUtilFactory.hpp"
