+++ include/okapi/impl/control/iterative/iterativeControllerFactory.hpp
@@ -8,9 +8,9 @@
 #ifndef _OKAPILIB_ITERATIVECONTROLLERFACTORY_HPP_
 #define _OKAPILIB_ITERATIVECONTROLLERFACTORY_HPP_
 
+#include "okapi/api/control/iterative/iterativePosPidController.hpp"
+#include "okapi/api/control/iterative/iterativeVelPidController.hpp"
 #include "okapi/api/util/mathUtil.hpp"
-#include "okapi/impl/control/iterative/iterativePosPidController.hpp"
-#include "okapi/impl/control/iterative/iterativeVelPidController.hpp"
 
 namespace okapi {
 class IterativeControllerFactory {
