+++ src/impl/control/async/asyncPosPidController.cpp
@@ -9,12 +9,6 @@
 #include "okapi/impl/util/timer.hpp"
 
 namespace okapi {
-AsyncPosPIDControllerArgs::AsyncPosPIDControllerArgs(std::shared_ptr<ControllerInput> iinput,
-                                                     std::shared_ptr<ControllerOutput> ioutput,
-                                                     const IterativePosPIDControllerArgs &iparams)
-  : input(iinput), output(ioutput), params(iparams) {
-}
-
 AsyncPosPIDController::AsyncPosPIDController(std::shared_ptr<ControllerInput> iinput,
                                              std::shared_ptr<ControllerOutput> ioutput,
                                              const double ikP, const double ikI, const double ikD,
@@ -22,15 +16,6 @@ AsyncPosPIDController::AsyncPosPIDController(std::shared_ptr<ControllerInput> ii
   : AsyncWrapper(iinput, ioutput,
                  std::make_unique<IterativePosPIDController>(
                    ikP, ikI, ikD, ikBias, std::make_unique<Timer>(),
-                   std::make_unique<SettledUtil>(std::make_unique<Timer>()))) {
-}
-
-AsyncPosPIDController::AsyncPosPIDController(std::shared_ptr<ControllerInput> iinput,
-                                             std::shared_ptr<ControllerOutput> ioutput,
-                                             const IterativePosPIDControllerArgs &iparams)
-  : AsyncWrapper(iinput, ioutput,
-                 std::make_unique<IterativePosPIDController>(
-                   iparams, std::make_unique<Timer>(),
                    std::make_unique<SettledUtil>(std::make_unique<Timer>()))) {
 }
 } // namespace okapi
