+++ src/impl/control/async/asyncVelPidController.cpp
@@ -10,13 +10,6 @@
 #include "okapi/impl/util/timer.hpp"
 
 namespace okapi {
-AsyncVelPIDControllerArgs::AsyncVelPIDControllerArgs(std::shared_ptr<ControllerInput> iinput,
-                                                     std::shared_ptr<ControllerOutput> ioutput,
-                                                     const IterativeVelPIDControllerArgs &iparams,
-                                                     const VelMathArgs &ivelMath)
-  : input(iinput), output(ioutput), params(iparams), velMath(ivelMath) {
-}
-
 AsyncVelPIDController::AsyncVelPIDController(std::shared_ptr<ControllerInput> iinput,
                                              std::shared_ptr<ControllerOutput> ioutput,
                                              const double ikP, const double ikD, const double ikF,
@@ -24,16 +17,6 @@ AsyncVelPIDController::AsyncVelPIDController(std::shared_ptr<ControllerInput> ii
   : AsyncWrapper(iinput, ioutput,
                  std::make_unique<IterativeVelPIDController>(
                    ikP, ikD, ikF, std::move(ivelMath), std::make_unique<Timer>(),
-                   std::make_unique<SettledUtil>(std::make_unique<Timer>()))) {
-}
-
-AsyncVelPIDController::AsyncVelPIDController(std::shared_ptr<ControllerInput> iinput,
-                                             std::shared_ptr<ControllerOutput> ioutput,
-                                             const IterativeVelPIDControllerArgs &iparams,
-                                             std::unique_ptr<VelMath> ivelMath)
-  : AsyncWrapper(iinput, ioutput,
-                 std::make_unique<IterativeVelPIDController>(
-                   iparams, std::make_unique<Timer>(),
                    std::make_unique<SettledUtil>(std::make_unique<Timer>()))) {
 }
 } // namespace okapi
